+++ Group_Project_PRG2782/frmStudents.cs
@@ -1,0 +1,25 @@
+﻿using System;
+using System.Collections.Generic;
+using System.ComponentModel;
+using System.Data;
+using System.Drawing;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using System.Windows.Forms;
+
+namespace Group_Project_PRG2782
+{
+    public partial class frmStudents : Form
+    {
+        public frmStudents()
+        {
+            InitializeComponent();
+        }
+
+        private void frmStudents_Load(object sender, EventArgs e)
+        {
+
+        }
+    }
+}
